+++ server/api.go
@@ -280,8 +280,8 @@ func (a *apiServer) SubscribeInternal(ctx context.Context, req *client.Subscribe
 		consumer = req.Consumer.ConsumerId
 	}
 	a.logger.Debugf("api: Subscribe "+
-		"[stream=%s, partition=%d, start=%s, offset=%d, timestamp=%d, group=%s, consumer=%s]",
-		req.Stream, req.Partition, req.StartPosition, req.StartOffset, req.StartTimestamp, group, consumer)
+		"[stream=%s, partition=%d, start=%s, offset=%d, timestamp=%d, group=%s, consumer=%s, reverse=%v]",
+		req.Stream, req.Partition, req.StartPosition, req.StartOffset, req.StartTimestamp, group, consumer, req.Reverse)
 
 	if group != "" && consumer == "" {
 		a.logger.Errorf("api: Failed to subscribe to partition: no consumer id provided")